refactor : email을 verify 패키지 안으로 이동
