+++ market/src/main/java/university/market/verify/email/controller/EmailController.java
@@ -1,4 +1,4 @@
-package university.market.email.controller;
+package university.market.verify.email.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import university.market.email.service.EmailService;
-import university.market.email.service.dto.CheckVerificationCodeRequest;
-import university.market.email.service.dto.EmailRequest;
+import university.market.verify.email.service.EmailService;
+import university.market.verify.email.service.dto.CheckVerificationCodeRequest;
+import university.market.verify.email.service.dto.EmailRequest;
 
 @RestController
 @RequestMapping("/api/email")
