+++ market/src/main/java/university/market/verify/email/domain/EmailVO.java
@@ -1,4 +1,4 @@
-package university.market.email.domain;
+package university.market.verify.email.domain;
 
 import java.sql.Timestamp;
 import lombok.Builder;
