+++ market/src/main/java/university/market/verify/email/service/dto/CheckVerificationCodeRequest.java
@@ -1,4 +1,4 @@
-package university.market.email.service.dto;
+package university.market.verify.email.service.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
