+++ market/src/main/resources/mapper/email-mapper.xml
@@ -3,9 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="university.market.email.mapper.EmailMapper">
+<mapper namespace="university.market.verify.email.mapper.EmailMapper">
 
-    <resultMap id="EmailResultMap" type="university.market.email.domain.EmailVO">
+    <resultMap id="EmailResultMap" type="university.market.verify.email.domain.EmailVO">
         <id property="id" column="id"/>
         <result property="email" column="email"/>
         <result property="verificationCode" column="verification_code"/>
@@ -13,7 +13,7 @@
         <result property="updatedAt" column="updated_at"/>
     </resultMap>
 
-    <insert id="saveVerificationCode" parameterType="university.market.email.domain.EmailVO">
+    <insert id="saveVerificationCode" parameterType="university.market.verify.email.domain.EmailVO">
         INSERT INTO email (email, verification_code, created_at, updated_at)
         VALUES (#{email}, #{verificationCode}, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
     </insert>
