+++ market/src/test/java/university/market/email/service/EmailServiceTest.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-import jakarta.mail.MessagingException;
 import java.lang.reflect.Field;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,14 +16,15 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
-import university.market.email.domain.EmailVO;
-import university.market.email.exception.EmailException;
-import university.market.email.exception.EmailExceptionType;
-import university.market.email.mapper.EmailMapper;
-import university.market.email.service.dto.CheckVerificationCodeRequest;
-import university.market.email.service.dto.EmailRequest;
-import university.market.email.utils.RandomUtil;
-import university.market.email.utils.content.EmailContent;
+import university.market.verify.email.domain.EmailVO;
+import university.market.verify.email.exception.EmailException;
+import university.market.verify.email.exception.EmailExceptionType;
+import university.market.verify.email.mapper.EmailMapper;
+import university.market.verify.email.service.EmailServiceImpl;
+import university.market.verify.email.service.dto.CheckVerificationCodeRequest;
+import university.market.verify.email.service.dto.EmailRequest;
+import university.market.verify.email.utils.RandomUtil;
+import university.market.verify.email.utils.content.EmailContent;
 
 import jakarta.mail.internet.MimeMessage;
