+++ market/src/test/java/university/market/email/util/EmailContentGenerateTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import university.market.email.utils.content.EmailContent;
+import university.market.verify.email.utils.content.EmailContent;
 
 @ExtendWith(MockitoExtension.class)
 public class EmailContentGenerateTest {
